Update README.md
### Jacob Praymore Dolapo

**CEO at Cobnet Maher Collection & CRM at Damako Gas Ltd**

Jacob Praymore Dolapo embodies the synergy between technology and healthcare, championing innovation and efficiency in both realms. With a background in customer relationship management and project planning, Jacob's leadership as CEO of Cobnet Maher Collection and Customer Relationship Manager at Damako Gas Ltd is characterized by a relentless pursuit of excellence.

Driven by a deep-rooted passion for leveraging technology to improve healthcare outcomes, Jacob is currently pursuing a B.Sc. in Healthcare Management. His journey is marked by a dedication to continuous learning and professional growth, evident in his certifications in Project Management, HSE, and various other fields.

As an active member of the Africa Society of Health Professionals (ASHP), Jacob is at the forefront of driving transformative change in healthcare delivery. His expertise in data analysis and technology integration empowers him to identify innovative solutions to complex challenges, ultimately enhancing patient care and operational efficiency.

Jacob's vision extends beyond the confines of traditional healthcare, as he explores emerging technologies and digital innovations to shape the future of the industry. His strategic mindset, coupled with a commitment to collaboration and excellence, positions him as a catalyst for progress at the intersection of technology and healthcare.

**Contact:**
- Email: praymorejacob@gmail.com
- LinkedIn: [Jacob Praymore Dolapo](https://www.linkedin.com/in/jacob-praymore-dolapo-573849250?utm_source=share&utm_campaign=share_via&utm_content=profile&utm_m)